Reworked RR implementation to fix bug
Initially this was going to be a small change, to fix some minor errors
in scheduling order and timing - but, it ended up requiring a full
rewrite (not modeled after any of the other mode functions), due to some
subtleties in the RR scheme. The best way ended up being to simulate the
actual ready queue used by a live scheduler.