Tasks and 1 dummy program that is called
So the comments on all these are pretty self explanatory. I didnt think
about how you would be using 1 and 2 to test 3, so I just hard coded the
memory, but you may want to add the ability to send in memory allocation
sizes in as a command line argument for ease of use in testing with
task3. Also, Task2a simply has the parent call 3 seperate children, in
the spec it doesnt say they have to alllocate memory (I'm guessing this
is because its a test to show what memory does simply on child creation,
i.e. does the child get any extra memory automatically). However, I am
not sure why it outputs the memory info for child 2 twice and never 3,
and then also never finishes executing. Perhaps you will see the
problem.